Moved cppNameStyle from CPlusPlus.ts to support/Strings.ts and renamed it to makeNameStyle. Added 2 new parameters to makeNameStyle:  - legalizeName: A language-specific function that legalizes a name to render in a given language  - prefix: An optional parameter that specifies a prefix to prepend to names
Moved NamingStyle enumerator from CPlusPlus.ts to support/Strings.ts to support the above change

CPlusPlus.ts:
 - Removed multiple private naming function variables and implemented them in their corresponding `makeXxxNamer` methods so they may be modified by inheritors
 - Changed the constructor to have C++ properties as default parameters instead of being hardcoded so that inheritors can make their language targetable by quicktype
 - updated code to use makeNamStyle instead of cppNameStyle
Removed private parameter naming functions

Fixed a typo in the constructor for the C++ target language and made it more clear that the prefix function will not in makeNameStyle if no prefix is given

Checked for if defined for the non-default behaviour for prefix checking in makeNameStyle
